third_party/blink/web_tests/external/wpt/html/semantics/popovers/popover-focus-5_commandforelement-expected.txt

This is a testharness.js-based test.
[FAIL] Focusing elements inside a popover with a details element inside, using commandforelement
  assert_equals: tab press should move to close popover button expected "close" but got "after"
[FAIL] Focusing elements inside a popover with a details element inside, and the after button adjacent to invoker, using commandforelement
  assert_equals: tab press should move to close popover button expected "close" but got "after"
[FAIL] Focusing elements inside a popover with a custom-element with delegatesfocus inside, using commandforelement
  assert_equals: tab press should move to close popover button expected "close" but got "after"
[FAIL] Focusing elements inside a popover with a custom-element with a slotted button inside, using commandforelement
  assert_equals: tab press should move to close popover button expected "close" but got "after"
[FAIL] Focusing elements inside a popover with custom-element with a slotted button, followed by a details element, where the after button is adjacent to the invoker, using commandforelement
  assert_equals: tab press should move to active element 2 expected "2" but got "after"
Harness: the test ran to completion.
